@@ -34,7 +34,10 @@ class RapidRecord extends BaseMongoModel
     /**
      * @inheritDoc
      */
-    protected $dates = ['created_at', 'updated_at'];
+    protected $casts = [
+        'created_at' => 'datetime',
+        'updated_at' => 'datetime'
+    ];
 
     /**
      * @var string
