@@ -12,7 +12,7 @@
     <div class="col-12 col-md-10 offset-md-1">
         <div class="card white box-shadow py-5 my-5 p-sm-5">
             <div class="col-6 mx-auto">
-                <form action="{{ route('app.password.email') }}" method="post" role="form">
+                <form action="{{ route('app.password.email') }}" method="post" role="form" class="recaptcha">
                     @csrf
                     <div class="form-group">
                         <label for="email" class="col-form-label required">{{ t('Email') }}:</label>
